[delete] : usecaseディレクトリを削除
ややこしいし、機能していないので削除する
