+++ src/hooks/index.ts
@@ -7,17 +7,17 @@ export { default as useFilteredNotes } from './useFilteredNotes';
 
 export { default as useAuth } from './useAuth';
 
-export { default as useNote } from './usecase/note/useNote';
-export { default as usePostNote } from './usecase/note/usePostNote';
-export { default as usePutNote } from './usecase/note/usePutNote';
-export { default as useDeleteNote } from './usecase/note/useDeleteNote';
-export { default as useNoteList } from './usecase/note/useNoteList';
+export { default as useNote } from './note/useNote';
+export { default as usePostNote } from './note/usePostNote';
+export { default as usePutNote } from './note/usePutNote';
+export { default as useDeleteNote } from './note/useDeleteNote';
+export { default as useNoteList } from './note/useNoteList';
 
-export { default as useTag } from './usecase/tag/useTag';
-export { default as usePostTag } from './usecase/tag/usePostTag';
-export { default as usePutTag } from './usecase/tag/usePutTag';
-export { default as useDeleteTag } from './usecase/tag/useDeleteTag';
-export { default as useTagList } from './usecase/tag/useTagList';
+export { default as useTag } from './tag/useTag';
+export { default as usePostTag } from './tag/usePostTag';
+export { default as usePutTag } from './tag/usePutTag';
+export { default as useDeleteTag } from './tag/useDeleteTag';
+export { default as useTagList } from './tag/useTagList';
 
-export { default as useUser } from './usecase/user/useUser';
-export { default as useDeleteUser } from './usecase/user/useDeleteUser';
+export { default as useUser } from './user/useUser';
+export { default as useDeleteUser } from './user/useDeleteUser';
