+++ src/hooks/tag/usePostTag.ts
@@ -1,10 +1,10 @@
 import { TagRepository } from '@/repositories';
 import { useState } from 'react';
-import useNotification from '../../useNotification';
+import useNotification from '../useNotification';
 import useTagList from './useTagList';
 import { useNavigate } from 'react-router-dom';
 
-export default function useDeleteTag(id: string) {
+export default function usePostTag() {
 	const [isLoading, setIsLoading] = useState<boolean>(false);
 	const [error, setError] = useState<Error | null>(null);
 	const tagRepository = new TagRepository();
@@ -13,26 +13,25 @@ export default function useDeleteTag(id: string) {
 	const { mutate: mutateTagList } = useTagList();
 	const Navigate = useNavigate();
 
-	// 削除の副作用
-	const effect = async () => {
+	const effect = async (id: string) => {
 		await mutateTagList();
-		Navigate('/app/tag');
+		Navigate(`/app/tag/${id}`);
 	};
 
-	const deleteTag = async () => {
+	const postTag = async (name: string) => {
 		setIsLoading(true);
 		try {
-			await tagRepository.delete(id);
+			const result = await tagRepository.post(name);
 			setError(null);
-			await effect();
+			await effect(result.id);
 			dispatchNotification({
 				severity: 'success',
-				message: '削除に成功しました。',
+				message: '送信に成功しました。',
 			});
 		} catch (e) {
 			dispatchNotification({
 				severity: 'error',
-				message: '削除に失敗しました。',
+				message: '送信に失敗しました。',
 			});
 			setError(e as Error);
 		} finally {
@@ -40,5 +39,5 @@ export default function useDeleteTag(id: string) {
 		}
 	};
 
-	return { deleteTag, isLoading, error };
+	return { postTag, isLoading, error };
 }
